Add stx-ceph-manager image in tag-mgmt
The app rook-ceph-migration uses stx-ceph-manager image,
however, the image is not tagged.

This commit adds the stx-ceph-manager in tag-mgmt according to
the rook-ceph-migration-helm (upstream) version from the commit
below:

https://review.opendev.org/c/starlingx/rook-ceph/+/898733

After this review is merged, a new review will be created in the
rook-ceph repository to adjust the use of the image with the new
tag.

Story: 2011055
Task: 49813

Change-Id: Ib163fdee84158899b45ffbc89028d05e6e164193
Signed-off-by: Gabriel de Araújo Cabral <gabriel.cabral@windriver.com>