debrepack: Fixed debver inconsistent issue
When python yaml module loads meta_data.yaml into a dict
with 'yaml.full_load', it discards trailing zeros in the
'debver' that has two digits before the decimal point,
and followed by a trailing zero after the decimal point.
For example, if the 'debver' in meta_data.yaml is '23.10',
the version in debrepack is '23.1' which will cause build
failure.

Test Plan:
Pass: The fix should not impact the normal build
      build-pkgs -a

Pass: Add new package 'mlnx-ofed-kernel_23.10' with 'debver'
      in meta_data.yaml is '23.10':
      build-pkgs -c -p mlnx-ofed-kernel

Closes-bug: 2060436

Change-Id: I33042f1d15440e72ad8a4b33061390f7dbb9bb64
Signed-off-by: Haiqing Bai <Haiqing.Bai@windriver.com>
[Cherry pick from https://review.opendev.org/c/starlingx/root/+/915251]
Signed-off-by: Jiping Ma <jiping.ma2@windriver.com>