Add logic for building extra helm build
This adds the build-extra-helm-charts script
that looks for *.helm.build build files. And
then runs the build command provided in the
required BUILD_COMMAND field. It is also
expected an OUTPUTS field on the helm.build
with a pattern indicating the path to
outputs of this build rlative to helm output

Testplan:
PASS - Run jenkins job to run custom helm app
       build

Story: 2011098
Task: 50092

Change-Id: I23cb4a488af1eaf83cfe5c0216557b183744de08
Signed-off-by: Romulo Leite <romulo.leite@windriver.com>