build-image: rename ISO to reflect build timestamp
Builds generated by Jenkins are identified by a timestamp, set in the
environment, such as TIMESTAMP=2024-06-12_19-00-08. Yet the ISO file
created by build-image is named after the clock value at the point of
creation, ie a different time stamp.

This patch renames the ISO to match the overall build timestamp, if
TIMESTAMP is set in the environment.

EXAMPLE
==========================
* Original file created by LAT SDK is similar to:
    starlingx-intel-x86-64-20240613180213-cd.iso
* With TIMESTAMP=2024-06-12_19-00-08 it's renamed to:
    starlingx-intel-x86-64-2024-06-12_19-00-08-cd.iso
  along with the signatures and symlinks

TESTS
==========================
* Call build-image with TIMESTAMP set and make sure the ISO file, the
  signature and the symlinks are generated correctly
* Call build-image w/o the TIMESTAMP and check that the files are not
  renamed

Story: 2010055
Task: 50358

Signed-off-by: Davlet Panech <davlet.panech@windriver.com>
Change-Id: I184f6a4f79688e0c8a3029d2aafa22c383b5a524