Add package cache update to downloader
On an existing workspace it is common to get
downloader errors if the packages cache has
not being updated for a while. To avoid this
the `downloader` will now run the `apt-get update`
command automatically.

Test plan:

PASS: Run downloader in a "old" environment
PASS: Run a fresh new environment downloader
PASS: Build-pkgs -a

Closes-Bug: 2039298

Change-Id: Icdaa942da231434d66fcd1a5ff4f89e3d164abfb
Signed-off-by: Romulo Leite <romulo.leite@windriver.com>