Add check for downloader source downloads
It was seen in [1] that Downloader was silently ignoring
errors while downloading source packages. Specifically in
packages that use `archive` in the meta_data.

This commit will add a handler for this failure by parsing
the package .dsc and checking if all necessary files were
correctly downloaded

Test plan:
PASS - downloader -s -b
PASS - serve a package with 1 .tar missing and notice failure

Closes-bug: 2039298

[1] https://bugs.launchpad.net/starlingx/+bug/2039298

Change-Id: I54bf8e41a5342f96a2ac86cf9587096ef0112e0c
Signed-off-by: Romulo Leite <romulo.leite@windriver.com>