Tox: Fix Semaphore object AttributeError
Some tests, such as sysinv-tox-py39 on the config repo,
are spamming this error message in the test logs [1]:

AttributeError: 'Semaphore' object has no attribute '_at_fork_reinit'

This issue is fixed on eventlet 0.30.0

In addition, eventlet 0.32.0 has a dnspython v2 compatibility fix.
Since dnspython 2.0.0 is the box version in bullseye, this update
is included as well.

Changes between versions 0.26.1 and 0.32.0 are few
and should not affect test results.

Ref:
[1] Logs from a config repo review:
https://00a1c49e3c0449afb12b-dfb7731ce6789292a31228a6fdf28206.ssl.cf1.rackcdn.com/923576/2/check/sysinv-tox-py39/507dc4f/job-output.txt
[2] Eventlet changelog:
https://github.com/eventlet/eventlet/blob/master/NEWS

Test Plan:
pass - On the config repo, run:
       tox -e py39 -c sysinv/sysinv/sysinv/tox.ini
       with UPPER_CONSTRAINTS_FILE env variable set
       to the updated upper-constraints.txt

Closes-Bug: 2064660

Change-Id: I2727b2f6bd955dcf9d917308f4795594f5a063e0
Signed-off-by: Leonardo Fagundes Luz Serrano <Leonardo.FagundesLuzSerrano@windriver.com>