Copy only the latest patch metadata
When creating a pre-patched ISO we need to move the metadata of the
patches we are applying to two separated folder: patches and upgrades.
This changes makes that only the metadata from the latest patch is
copied to both folders. Also the metadata in upgrades folder will
receive required information about the ostree repository inside the ISO.

Test plan:
    PASS: Create pre-patched ISO, mount it, check if metadata files are
    located in the correct folders and are valid XML.
    PASS: Execute full install AIO-SX, check `software list` command if
    only the latest patch is displayed.
    PASS: Execute "sofware upload <prepatched iso> <sig file>" and check
    if only the latest patch is displayed.

Story: 2010676
Task: 50726

Change-Id: I121a4fb688022bfb190bdf06669233e8e2f62fe7
Signed-off-by: Dostoievski Batista <dostoievski.albinobatista@windriver.com>