Changing commit ostree check error to warning
The ostree_repo will not hold the entire history of commits every run.
Making the verification we do between the commits hashs inside
ostree_repo and the commits hashs inside the patch metadata not always
valid. This change the script to warn the user about this difference
instead of just stopping with an error.

Test plan:
    PASS: Build pre-patched ISO without the complete commits history in
        ostree_repo
    PASS: Build pre-patched ISO with the complete commits history inside
        ostree_repo

Story: 2010055
Task: 50835

Change-Id: Iab6fc966d095b0ec25ffe33d867928b1e424b44f
Signed-off-by: Dostoievski Batista <dostoievski.albinobatista@windriver.com>