Add --ignore flag to build-helm-charts.sh file
Recently a review was merged changing the cp command to rsync, so
the helm-toolkit Helm-chart would not be added to the final application
tarball.

https://review.opendev.org/c/starlingx/root/+/927047

This change is a follow-up to that bug-fix. The objective here is to
remove the hard coded helm-toolkit name from the rsync, and replace it
with a variable containing the contents of a new flag (--ignore).
This flag is to list all Helm-charts that are not to be included in
the final application-tarball.

Test-Plan:
PASS - Application tarball built without the charts defined by the flag
PASS - Application tarball is built if the flag is not used
PASS - STX-Openstack upload, apply, remove, delete and re-upload

Closes-Bug: 2077759

Change-Id: I3d04479c8fad108e8777368d28acee84b9f8f7a9
Signed-off-by: Daniel Caires <DanielMarques.Caires@windriver.com>