Add option to use tmpfs during build-pkgs
Building in a tmpfs speeds the build process and greatly reduces
the disk io required during a build, leaving the host far more
responsive to other tasks.

Added option --tmpfs_percentage=N to the build-pkgs command
to control how much memory can be used for tmpfs build environments.
The valid range is 0-50%. The default is 0% resulting in a traditional
disk only build.

The available memory on the host is multiplied by the percentage
to determine how much is availabe for tmpfs in total.  This total
is then assigned to up to N-1 of the N parallel build environments,
with the minimum tmpfs environment size being 10GB.
One environment is reserved for disk based builds.  First time
package builds are assigned to tmpfs based build environments in
preference over the disk base environments. However, if a package
has failed a prior build attempt, subsequents attemps will only
occure on a disk based build environmnet.  This may help if the
build failed due to a too small tmpfs build environment.  It also
may leave the environment intact for mor indepth debugging.

Testing has revealed that 20gb is required to build the largest
packages (linux, ceph, kubernetes).  In order to avoid costly
rebuilds of these large packages, the choice of tmpfs percentage
and number of parallel builds is important.

e.g. on a single user host with 128GB of memory, an appropriate
choice might be...

    --parallel=4 --tmpfs_percentage=50

...yielding 3 tmpfs build environments of aprox 20GB each.
A higher parallelism, or a lower tempfs percentage will result
in build environments that drop below 20GB, and you might start
seeing rebuilds of large packages.

An advisory that a package requires a specified amount of space to
build successfully can be added by adding 'XS-Build-Size: <size>'
to the isource section of the debian 'control' file.
e.g. 'XS-Build-Size: 15GB'

Testing:
   - build-pkgs without --tmpfs_percentage
     Only disk base build envoronments are used.
   - build-pkgs --parallel=4 --tmpfs_percentage=20
     On 128GB machine, Only 2 10-12 GB tmpfs are used,
     the other two remanin disk based.
     Large packages fail in 10GB tmpfs, and a pass when retried on disk.
   - build-pkgs --parallel=4 --tmpfs_percentage=50
     On 128GB machine, Creates 3 18-20 GB tmpfs.
     Large packages build in tmpfs without need for rebuild.

Partial-Bug: 2081843
Depends-On: https://review.opendev.org/c/starlingx/tools/+/930348
Change-Id: I976be47a62c1fb3f319c228b32e5b4f02863df78
Signed-off-by: Scott Little <scott.little@windriver.com>