Zuul: 'stevedore' compatibility with setuptools/packaging
Setuptools >71 is installed in the zuul env as a setup
before running the test jobs. Recent versions of setuptools
have packaging >22 as a requirement. This constraint
was updated in [1]. Stevedore <= 4.8.0 is incompatible
with these updates [2].

In this commit, changed stevedore upper version constraint
in zuul tests to 5.0.0, matching [3].

Test Plan:
pass - Mock review to update repo [4], where the constraint
       was causing stx-software-tox-py39 zuul job to fail
pass - On the config repo, ran sysinv-tox-py39 zuul job
       locally, since stevedore is cited in
       sysinv/sysinv/sysinv/sysinv/common/utils.py

Refs:
[1] Updated 'packaging' python module version constraint
    https://review.opendev.org/c/starlingx/root/+/929167
[2] Forum post citing error message from the bug report
    https://github.com/python/importlib_metadata/issues/348
[3] STX Openstack stevedore constraint is set to 5.0.0
    https://opendev.org/starlingx/root/commit/0f050b28f2ae88e72154ba02efddb051f331aa61
[4] Mock review to the update repo to test change
    https://review.opendev.org/c/starlingx/update/+/929287

Closes-Bug: 2080674

Change-Id: I47ec888e8e56bcbbe547d323add9351bc540a89c
Signed-off-by: Leonardo Fagundes Luz Serrano <Leonardo.FagundesLuzSerrano@windriver.com>