Fetch the correct package instead of the first one
When running the patch-builder command with a recipe that includes
packages with similar names to other packages in the base-bullseye.lst,
the code would fetch the first one instead of the correct one. This
change verifies if the package name matches with the package specified
in the patch recipe bofore adding it to the patch.

Test plan:
    PASS - Run patch-builder with packages `gcc` and `cpp`, check if
    correct packages are inside the patch.
    PASS - Run patch-builder with all binaries packages on base-bullseye
    and check if correct packages are inside the patch.
    PASS - Run patch-builder with non-existent package and check if
    warning is shown.

Closes-bug: 2085579

Change-Id: Iee9de9db4ffa79fcfd3bb925258497ea98755770
Signed-off-by: Dostoievski Batista <dostoievski.albinobatista@windriver.com>