create-prepatched-iso: add secondary input ISO
Added a feature where the user may provide a secondary ISO
along with a list of ISO contents which should be taken
from this second ISO instead of the first.

For instance, if all output ISO contents should be taken from
the first ISO, but a folder named "kickstart" must be taken from
a second ISO, the user may run:

<script> -i <iso_A> -si <iso_B> -sc kickstart -p <patch> -o <iso_out>

In addition, some extensive re-factoring has been made on the
majority of the script now that the requirements for it
have stabilized. More logging has been added as well.

Test Plan:

pass: Generated ISO using as inputs:
      - Only a base ISO and a patch
      - Added a secondary ISO and secondary contents,
        and confirmed output ISO content matches selection
      - Custom base ostree repo

pass: Dummy jenkins pipeline
pass: Install and Unlock with output ISO

Story: 2011318
Task: 51505

Change-Id: I6ecc6eab3149ca1d05d53ac3591f5a96abca5270
Signed-off-by: Leonardo Fagundes Luz Serrano <Leonardo.FagundesLuzSerrano@windriver.com>