Set build order for openstackclients
Changes in the dependencies defined in the
upstream client packages may have created
a new circular dependency between the novaclient
and openstackclient packages.

This change is necessary to allow openstack to be
built without circular dependency errors with the
novaclients and openstackclients.

This change is needed to support:

https://review.opendev.org/c/starlingx/openstack-armada-app/+/937801

Test plan:

PASS build-pkgs -c -l openstack

Closes-Bug: 2092540

Change-Id: I1d41d1445c6a574e81b3974c821b14b2c3d2e3a4
Signed-off-by: vrochalo <vinicius.rochalobo@windriver.com>