Make builder nice to it's host
StarlingX build tools consume excessive cpu and io which can starve
non-build services, including kubernetes and calico. This results in
timeouts, restarts, and general instability of calico and kubernetes
which in turn can the build to fail. This affect is observed on the
main StarlingX build server.

The major build steps should employ 'nice' and 'ionice' to ensure
that critical services on the build host are not starved for cpu or io.

Depends-On: https://review.opendev.org/c/starlingx/tools/+/939821
Closes-bug: 2095512
Change-Id: Ib54076e796e4cd575bd6fb8e89b70dba8649c4d4
Signed-off-by: Scott Little <scott.little@windriver.com>
(cherry picked from commit 7715e8bef2f8af1a48a7740055bacf0abae3df51)