fetch_debs.py doesn't select packages correctly
When fetching the debs to be part of the patch we filter them based on
debian_iso_image.inc. That is not always correct as some packages may
need to be part of the patch but is not part of any
debian_iso_image.inc. This change allow us to add packages to the patch
that are explicitly listed in the patch recipe.

Test plan:
    PASS - Build patch with packages not present in debian_iso_image.inc
    PASS - Build patch with packages present in debian_iso_image.inc

Closes-Bug: 2098403

Change-Id: I1825f38190746f0c0743582e5df56d6479d0c8b2
Signed-off-by: Dostoievski Batista <dostoievski.albinobatista@windriver.com>