patch-iso-debian: use mount for extracting ISO
In a previous commit [1], the script was changed to use 7z
for extracting contents from the input ISO. This was done
because the main alternative was using 'mount', which
requires sudo privileges.

Unfortunately, 7z does not handle ownership and permissions [2],
and keeping them unchanged is required.

This commit changes the script back to using 'mount'
for extracting input ISO contents. This also means running it
now requires sudo privileges.

Ref:
[1] https://opendev.org/starlingx/root/commit/4f69113d93fbe65257b0e3ba8022fa2f1b71ccb0
[2] https://linux.die.net/man/1/7z

Test Plan:
pass - Generate a pre-patched ISO and check that the permissions
       on the output iso are the same as the ones in the input.

Closes-Bug: 2098385

Change-Id: I1e5e1f8ecdb94cafb577dcfe3651a1717abe19c7
Signed-off-by: Leonardo Fagundes Luz Serrano <Leonardo.FagundesLuzSerrano@windriver.com>