downloader: handle non-standard DEB file names
Binary package lists [1] currently support arbitrary URLs for the
packages we need to download, but there's a problem in the downloader
in that it assumes that basename of such URLs follows a standard naming
convention, ie <name>_<version>_<arch>.deb.

This fix allows us to specify any file name in such URLs. Downloader
will now save the deb file using the above convention, rregardles of how
the file name is spelled in the original source URL.

[1] https://opendev.org/starlingx/tools/src/branch/master/debian-mirror-tools/config/debian/common/base-bullseye.lst

TESTS
========================
* Clean mirrors/ directory, run downloader with and without this patch.
  Make sure it produces the same files before and after.
* Add this new package to base-bullseye.lst:
    bao 2.1.0 https://github.com/openbao/openbao/releases/download/v2.1.0/bao_2.1.0_linux_amd64.deb
* Re-run downloader (it fails due to file name discrepancy)
* Re-run downloader with this patch andm make sure the file is
  processed correctly

Story: 2010055
Task: 51735

Signed-off-by: Davlet Panech <davlet.panech@windriver.com>
Change-Id: I4cd31e065e371545052dd3c5ed20779445a8f004