patch-builder: Flag for adding precheck scripts
Currently, if the patching framework deb package (software.deb)
is one of the pkgs selected for the patch, patch-builder will
also extract a couple scripts (refered to here as 'precheck scripts')
and put them in the separatly in the patch for easier access.

The issue is that, if we create a patch that delivers an ostree repo
instead of a set of deb packages, then auto-inclusion of the
precheck scripts will never happen.

Added a Y/N flag the user can add in the patch XML (similar to the
reboot_required flag) to request inclusion of the precheck scripts.

Test Plan:
pass - flag is not mandatory
pass - if 'Y' regardless of debs requested, scripts are included
pass - if 'N', scripts are included only if software.deb is requested

Story: 2011498
Task: 52900

Change-Id: Ic9866bcd251bfe6f27ce4b60be604beccb815183
Signed-off-by: Leonardo Fagundes Luz Serrano <Leonardo.FagundesLuzSerrano@windriver.com>