Removed XML encoding from prepatched metadata
The change in [1] introduced an XML header into the ISO metadata
which includes XML version and encoding.

The issue is that current XML parsing on the STX runtime side
already decodes the file into a string assuming UTF-8 encoding
and, when the string is provided to the lxml module with an
encoding declaration, that triggers an error.

The solution is simply to not declare any encoding for the time being.

Test Plan:
pass: Parse metadata using lxml.etree.fromstring()

Refs:
[1] https://opendev.org/starlingx/root/commit/c8826f94da9391ddf6b4973ca4d74839a114dcd3

Closes-Bug: 2129132

Change-Id: I17369bb1340e8aa11c458c01eaedf8556c90a921
Signed-off-by: Leonardo Fagundes Luz Serrano <Leonardo.FagundesLuzSerrano@windriver.com>