Fix apt cache used by the --reuse code
Work around a serious bug in the newer upstream python apt implementation
that breaks when a rootdir != / is used in conjunction with the update method.
This unresolved bug describes the issue...
   Bug #1875601: python-apt update() ignores rootdir for cleanup
   https://bugs.launchpad.net/ubuntu/+source/python-apt/+bug/1875601

Extensive config overrides are required to get python apt to use the rootdir
correctly.  In the case of the update method, there is no workaround but to
call 'apt update' outside of python and with sudo.

This fix applies to the '--reuse' subsystem which uses an independent apt cache
from that used by the rest of package builder and delivered in
git SHA f51fadda64e7111d0aaf737b30abab8fff5b04d4.

Hopefully the upstream bug is resolved soon, and we can remove this ugly workaround.

Story: 2011360
Task: 52884
Change-Id: I7767ca3387b5f163f4b6f458807e44e374d656c3
Signed-off-by: Scott Little <scott.little@windriver.com>