Add stx-openbao-manager to tag-management list
The stx-openbao-manager docker image used to manage openbao application
is reinstated, and it will be required for the upcoming changes to the
openbao-manager helm chart. Adding the image to the image-tags.yaml, so
that it can be pulled with a proper tag.

Test Plan:
PASS    Image builds with the changes in src_ref
PASS    App lifecycle test (upload/apply/remove/delete) with the new
image

Story: 2011244
Task: 53382

Change-Id: Ibaeea8d054f144d1bafc3e901c400c380ef10385
Signed-off-by: Tae Park <tae.park@windriver.com>