Update stx-openbao-manager tag
Updating the stx-openbao-manager tag to stx.12.0-v1.32.10-1.
This is to include the updated version of openbao-monitor package used
in this image. The changes in that package is in this review:
https://review.opendev.org/c/starlingx/utilities/+/970868

Test Plan:
PASS    Image has the new openbao-monitor package
PASS    App lifecycle test (upload/apply/remove/delete) with the new
image
PASS    Baomon init/unseal and openbao secret store is successful

Story: 2011244
Task: 53524

Change-Id: Iba476a7f19ac41d3cf721b2d2ba6a61365aea873
Signed-off-by: Tae Park <tae.park@windriver.com>