create-prepatched-iso: Accept empty patch
This is an extension of the work done in [1], in which
the patch-builder was adjusted to accept making patches
with no deb pkgs in it.

In this commit, the script for creating pre-patched ISOs
is adjusted to also accept these empty patches.

Ref:
[1] https://review.opendev.org/c/starlingx/root/+/956588

Test Plan:
pass: Create prepatched ISO with empty patch as input

Story: 2011498
Task: 53541

Change-Id: If6cc07024f6c9f2af9c612b7bad8dee8042558a2
Signed-off-by: Leonardo Fagundes Luz Serrano <Leonardo.FagundesLuzSerrano@windriver.com>